// web-front/src/Types/resposta.ts
import React from 'react';

type resposta = {
    id: number;
    email: string;
    data_hora: Date;
    campus: string;
    modalidade: string;
    periodo: string;
    f: string;
    g: string;
    h: string;
    i: string;
    j: string;
    k: string;
    l: string;
    m: string;
    n: string;
    o: string;
    p: string;
    q: string;
    r: string;
    s: string;
    t: string;
}

export default resposta;